pyproject.toml
[project]
name = "nancyai"
version = "2.0"
description = "Telegram AI chat bot using opensource Llama."
authors = [
    {name = "JeevabharathiRK"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "python-dotenv>=1.0.0",
    "aiogram>=3.22.0",
    "groq>=0.8.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
nancy = "nancyai.bot:main"